requirements.txt
# File: requirements.txt
################################################################################
# File: requirements.txt
################################################################################
# Python dependencies for Misophonia Research RAG System

# Core dependencies
firebase-admin>=6.2.0
openai>=1.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
flask>=2.0.0
requests>=2.28.0

# PDF processing
PyPDF2>=3.0.0
unstructured>=0.10.0
pdfminer.six>=20221105

# Utilities
tqdm>=4.65.0
tabulate>=0.9.0
colorama>=0.4.6
argparse>=1.4.0

# Firebase/Google Cloud
google-cloud-firestore>=2.11.0

# For concurrent processing
concurrent-log-handler>=0.9.20

# Supabase integration
supabase~=2.0.0
